igni/requirements.txt
jinja2
pwntools
pyelftools
monkeyhex
